\name{as.grain}
\alias{as.grain}
\title{Conversion into \code{grain} class}
\description{Convert an object of class \code{cibn} into an object of class \code{grain}.}
\usage{as.grain(x)}
\arguments{
  \item{x}{An object of class \code{cibn}.}
}
\value{An object of class \code{grain}.}
%\details{}
%\references{}
\seealso{\link{new.cibn}.}
\examples{
## NOT RUN:
#bankrisk_bn <- new.cibn(bankrisk_code)
#G <- as.grain(bankrisk_bn)
#G
}